Move travis to staged build
This adds stages to the travis build. This setting here allows travis to
only run the unit-tests with different php-versions  after the code has been validated wth one php-version.

It also uses the PHAR-file of the coveralls-tool to upload coverage to
coveralls.io only when a push to master has been done. Before it would
have pushed on every build even for a PR which might not be what is
wanted...